Provided values for charset/line endings
The default is to use the host OS' native/preferred format.
I set it to UTF8 with CRLF, so it will use the same on all platforms.
CRLF is used because notepad.exe is too stupid for LF-only, and i often
show people source code on computers without a decent text editor.